feat: modularize extractors and resolvers out of parser.py and graph.py (Phase 2)
Extract per-language logic into dedicated packages:
- extractors/ — visibility, signatures, docstrings, bindings, heritage
- resolvers/ — Python, TS/JS, Go, Rust, C/C++, generic stem fallback
- framework_edges.py — Django, FastAPI, Flask, pytest conftest detection

parser.py drops from 1,806 to 796 lines (pure orchestration).
graph.py drops from 1,286 to 646 lines. Delete dead parsers/ stubs.
