--- packages/core/src/repowise/core/ingestion/parsers/__init__.py
@@ -1,4 +1,0 @@
-# This directory is intentionally empty.
-# Language-specific parsing is handled by the unified ASTParser in
-# repowise.core.ingestion.parser using per-language .scm query files
-# in packages/core/queries/.
