--- packages/core/src/repowise/core/ingestion/parsers/base.py
@@ -1,3 +1,0 @@
-# Replaced by the unified ASTParser architecture.
-# See packages/core/src/repowise/core/ingestion/parser.py and
-# packages/core/queries/*.scm for the current implementation.
