--- packages/core/src/repowise/core/ingestion/parsers/python.py
@@ -1,2 +1,0 @@
-# Replaced by the unified ASTParser architecture.
-# See packages/core/queries/python.scm for the symbol/import queries.
