--- packages/core/src/repowise/core/ingestion/parsers/typescript.py
@@ -1,2 +1,0 @@
-# Replaced by the unified ASTParser architecture.
-# See packages/core/queries/typescript.scm and javascript.scm.
